+++ visualization/app/codeCharta/codeCharta.module.ts
@@ -8,7 +8,8 @@ import "./state/state.module"
 
 import { codeChartaComponent } from "./codeCharta.component"
 import { CodeChartaService } from "./codeCharta.service"
-import { CodeChartaMouseEventService } from "./codeCharta.mouseEvent.service"
+import { downgradeComponent } from "@angular/upgrade/static"
+import { LoadingFileProgressSpinnerComponent } from "./ui/loadingFileProgressSpinner/loadingFileProgressSpinner.component"
 
 angular.module("app.codeCharta", ["app.codeCharta.state", "app.codeCharta.ui"])
 
@@ -16,4 +17,4 @@ angular
 	.module("app.codeCharta")
 	.component(codeChartaComponent.selector, codeChartaComponent)
 	.service(camelCase(CodeChartaService.name), CodeChartaService)
-	.service(camelCase(CodeChartaMouseEventService.name), CodeChartaMouseEventService)
+	.directive("ccLoadingFileProgressSpinner", downgradeComponent({ component: LoadingFileProgressSpinnerComponent }))