+++ visualization/app/codeCharta/state/store/store.ts
@@ -1,4 +1,5 @@
 import { Action, createStore } from "redux"
+import { EffectsModule } from "../angular-redux/effects/effects.module"
 
 import rootReducer from "./state.reducer"
 
@@ -14,10 +15,19 @@ export class Store {
 
 	private static initialize() {
 		Store._store = Store.createStore()
+		const originalDispatch = Store._store.dispatch
+		// @ts-ignore
+		Store._store.dispatch = function (action: Action) {
+			originalDispatch(action)
+			EffectsModule.actions$.next(action)
+			return action
+		}
 	}
 
 	static get store() {
-		if (!Store._store) Store.initialize()
+		if (!Store._store) {
+			Store.initialize()
+		}
 		return Store._store
 	}