+++ visualization/app/codeCharta/ui/fileChooser/fileChooser.component.ts
@@ -8,6 +8,8 @@ import { ExportCCFile } from "../../codeCharta.api.model"
 import zlib from "zlib"
 import { CUSTOM_CONFIG_FILE_EXTENSION, CustomConfigHelper } from "../../util/customConfigHelper"
 import { getCCFileAndDecorateFileChecksum } from "../../util/fileHelper"
+import { parseGameObjectsFile } from "../../util/gameObjectsParser/gameObjectsImporter"
+import { validateGameObjects } from "../../util/gameObjectsParser/gameObjectsValidator"
 
 export class FileChooserController {
 	private files: NameDataPair[] = []
@@ -38,7 +40,11 @@ export class FileChooserController {
 				}
 
 				reader.onload = event => {
-					content = (isCompressed ? zlib.unzipSync(Buffer.from(<string>event.target.result)) : event.target.result).toString()
+					const result = event.target.result.toString()
+					content = isCompressed ? zlib.unzipSync(Buffer.from(<string>event.target.result)).toString() : result
+					if (result.includes("gameObjectPositions") && validateGameObjects(result)) {
+						content = JSON.stringify(parseGameObjectsFile(result))
+					}
 				}
 
 				reader.onloadend = () => {