+++ visualization/app/codeCharta/ui/ribbonBar/ribbonBar.e2e.ts
@@ -2,11 +2,13 @@ import { goto } from "../../../puppeteer.helper"
 import { RibbonBarPageObject } from "./ribbonBar.po"
 import { SearchPanelPageObject } from "../searchPanel/searchPanel.po"
 import { AreaSettingsPanelPageObject } from "../areaSettingsPanel/areaSettingsPanel.po"
+import { FileChooserPageObject } from "../fileChooser/fileChooser.po"
 
 //Commented out flaky test
 
 describe("RibbonBar", () => {
 	let searchPanel: SearchPanelPageObject
+	let fileChooser: FileChooserPageObject
 	//let searchPanelModeSelector: SearchPanelModeSelectorPageObject
 	let ribbonBar: RibbonBarPageObject
 	//let metricChooser: MetricChooserPageObject
@@ -14,6 +16,7 @@ describe("RibbonBar", () => {
 
 	beforeEach(async () => {
 		searchPanel = new SearchPanelPageObject()
+		fileChooser = new FileChooserPageObject()
 		//searchPanelModeSelector = new SearchPanelModeSelectorPageObject()
 		ribbonBar = new RibbonBarPageObject()
 		//metricChooser = new MetricChooserPageObject()
@@ -30,21 +33,6 @@ describe("RibbonBar", () => {
 		const actual = await metricChooser.getAreaMetricValue()
 		expect(actual).toContain("600")
 	})*/
-
-	it("focus of ui element should be removed on ribbonBar toggle", async () => {
-		const panel = "color-metric"
-		let isColorSettingsPanelOpen = await ribbonBar.togglePanel(panel)
-		expect(isColorSettingsPanelOpen).toBeTruthy()
-		await ribbonBar.focusSomething()
-		const activeBefore = await ribbonBar.getActiveClassName()
-
-		isColorSettingsPanelOpen = await ribbonBar.togglePanel(panel)
-		expect(isColorSettingsPanelOpen).toBeFalsy()
-
-		const activeAfter = await ribbonBar.getActiveClassName()
-		expect(activeBefore).not.toBe("ng-scope")
-		expect(activeAfter).toBe("ng-scope")
-	})
 
 	describe("opening and closing ribbon-bar cards", () => {
 		it("search-panel-card", async () => {
@@ -124,5 +112,19 @@ describe("RibbonBar", () => {
 		expect(await AreaSettingsPanelPageObject.toggleDefaultMargin()).toBeFalsy()
 
 		expect(await ribbonBar.isPanelOpen(areaPanel)).toBeTruthy()
+	})
+
+	it("should not find edge-metric-panel when file has no edge metrics", async () => {
+		const edgePanel = "edge-metric"
+		await fileChooser.openFiles(["./app/codeCharta/assets/sample3.cc.json"])
+
+		expect(await ribbonBar.isElementPresent(edgePanel)).toBeFalsy()
+	})
+
+	it("should find edge-metric-panel when file has edge metrics", async () => {
+		const edgePanel = "edge-metric"
+		await fileChooser.openFiles(["./app/codeCharta/assets/sample1.cc.json"])
+
+		expect(await ribbonBar.isElementPresent(edgePanel)).toBeTruthy()
 	})
 })