+++ visualization/app/codeCharta/ui/searchPanel/searchBar/selectors/isSearchPatternEmpty.selector.ts
@@ -1,0 +1,6 @@
+import { createSelector } from "../../../../state/angular-redux/createSelector"
+import { searchPatternSelector } from "../../../../state/store/dynamicSettings/searchPattern/searchPattern.selector"
+
+export const _isSearchPatternEmpty = (searchPattern: string) => searchPattern === "" || searchPattern === "!" || searchPattern === ","
+
+export const isSearchPatternEmptySelector = createSelector([searchPatternSelector], _isSearchPatternEmpty)