+++ visualization/app/codeCharta/ui/searchPanelModeSelector/searchPanelModeSelector.component.spec.ts
@@ -2,15 +2,12 @@ import "./searchPanelModeSelector.module"
 import { SearchPanelModeSelectorController } from "./searchPanelModeSelector.component"
 import { instantiateModule, getService } from "../../../../mocks/ng.mockhelper"
 import { IRootScopeService } from "angular"
-import { SearchPanelMode, BlacklistType } from "../../codeCharta.model"
+import { BlacklistType } from "../../codeCharta.model"
 import { BlacklistService } from "../../state/store/fileSettings/blacklist/blacklist.service"
-import { StoreService } from "../../state/store.service"
-import { SearchPanelModeService } from "../../state/store/appSettings/searchPanelMode/searchPanelMode.service"
 
 describe("SearchPanelModeSelectorController", () => {
 	let searchPanelModeSelectorController: SearchPanelModeSelectorController
 	let $rootScope: IRootScopeService
-	let storeService: StoreService
 
 	beforeEach(() => {
 		restartSystem()
@@ -21,11 +18,10 @@ describe("SearchPanelModeSelectorController", () => {
 		instantiateModule("app.codeCharta.ui.searchPanelModeSelector")
 
 		$rootScope = getService<IRootScopeService>("$rootScope")
-		storeService = getService<StoreService>("storeService")
 	}
 
 	function rebuildController() {
-		searchPanelModeSelectorController = new SearchPanelModeSelectorController($rootScope, storeService)
+		searchPanelModeSelectorController = new SearchPanelModeSelectorController($rootScope)
 	}
 
 	describe("constructor", () => {
@@ -35,24 +31,6 @@ describe("SearchPanelModeSelectorController", () => {
 			rebuildController()
 
 			expect(BlacklistService.subscribe).toHaveBeenCalledWith($rootScope, searchPanelModeSelectorController)
-		})
-
-		it("should subscribe to SearchPanelService", () => {
-			SearchPanelModeService.subscribe = jest.fn()
-
-			rebuildController()
-
-			expect(SearchPanelModeService.subscribe).toHaveBeenCalledWith($rootScope, searchPanelModeSelectorController)
-		})
-	})
-
-	describe("onSearchPanelModeChanged", () => {
-		it("should update searchPanelMode", () => {
-			const searchPanelMode = SearchPanelMode.blacklist
-
-			searchPanelModeSelectorController.onSearchPanelModeChanged(searchPanelMode)
-
-			expect(searchPanelModeSelectorController["_viewModel"].searchPanelMode).toEqual(SearchPanelMode.blacklist)
 		})
 	})
 
@@ -71,22 +49,6 @@ describe("SearchPanelModeSelectorController", () => {
 
 			expect(searchPanelModeSelectorController["_viewModel"].flattenListLength).toEqual(1)
 			expect(searchPanelModeSelectorController["_viewModel"].excludeListLength).toEqual(2)
-		})
-	})
-
-	describe("onToggleSearchPanelMode", () => {
-		it("should select if not already selected", () => {
-			searchPanelModeSelectorController.onToggleSearchPanelMode(SearchPanelMode.treeView)
-
-			expect(storeService.getState().appSettings.searchPanelMode).toEqual(SearchPanelMode.treeView)
-		})
-
-		it("should unselect if already selected", () => {
-			searchPanelModeSelectorController["_viewModel"].searchPanelMode = SearchPanelMode.treeView
-
-			searchPanelModeSelectorController.onToggleSearchPanelMode(SearchPanelMode.treeView)
-
-			expect(storeService.getState().appSettings.searchPanelMode).toEqual(SearchPanelMode.minimized)
 		})
 	})
 })