+++ visualization/app/codeCharta/ui/ui.ts
@@ -14,7 +14,6 @@ import "./presentationModeButton/presentationModeButton.module"
 import "./centerMapButton/centerMapButton.module"
 import "./searchPanel/searchPanel.module"
 import "./searchPanelModeSelector/searchPanelModeSelector.module"
-import "./searchBar/searchBar.module"
 import "./areaSettingsPanel/areaSettingsPanel.module"
 import "./blacklistPanel/blacklistPanel.module"
 import "./codeMap/codeMap.module"
@@ -23,7 +22,6 @@ import "./dialog/dialog.module"
 import "./fileChooser/fileChooser.module"
 import "./fileExtensionBar/fileExtensionBar.module"
 import "./heightSettingsPanel/heightSettingsPanel.module"
-import "./loadingGif/loadingGif.module"
 import "./metricChooser/metricChooser.module"
 import "./metricType/metricType.module"
 import "./rangeSlider/rangeSlider.module"
@@ -60,7 +58,6 @@ angular
 		"app.codeCharta.ui.centerMapButton",
 		"app.codeCharta.ui.searchPanel",
 		"app.codeCharta.ui.searchPanelModeSelector",
-		"app.codeCharta.ui.searchBar",
 		"app.codeCharta.ui.metricType",
 		"app.codeCharta.ui.areaSettingsPanel",
 		"app.codeCharta.ui.blacklistPanel",
@@ -70,7 +67,6 @@ angular
 		"app.codeCharta.ui.fileChooser",
 		"app.codeCharta.ui.fileExtensionBar",
 		"app.codeCharta.ui.heightSettingsPanel",
-		"app.codeCharta.ui.loadingGif",
 		"app.codeCharta.ui.metricChooser",
 		"app.codeCharta.ui.metricType",
 		"app.codeCharta.ui.rangeSlider",