+++ visualization/app/material/material.module.ts
@@ -13,6 +13,7 @@ import { MatCheckboxModule } from "@angular/material/checkbox"
 import { MatToolbarModule } from "@angular/material/toolbar"
 import { MatFormFieldModule } from "@angular/material/form-field"
 import { MatInputModule } from "@angular/material/input"
+import { MatProgressSpinnerModule } from "@angular/material/progress-spinner"
 
 const materialModules = [
 	MatSelectModule,
@@ -28,7 +29,8 @@ const materialModules = [
 	MatCheckboxModule,
 	MatToolbarModule,
 	MatFormFieldModule,
-	MatInputModule
+	MatInputModule,
+	MatProgressSpinnerModule
 ]
 
 @NgModule({